# docs/swagger.yaml
definitions:
...
  models.AchievementPeriodStat:
    properties:
      period:
        description: 'contoh: 2024-01'
        type: string
      total:
        type: integer
    type: object
...
  models.AchievementStatisticsResponse:
    properties:
      byPeriod:
        items:
          $ref: '#/definitions/models.AchievementPeriodStat'
        type: array
      byType:
        items:
          $ref: '#/definitions/models.AchievementTypeStat'
        type: array
      competitionLevels:
        items:
          $ref: '#/definitions/models.CompetitionLevelStat'
        type: array
      topStudents:
        items:
          $ref: '#/definitions/models.TopStudentStat'
        type: array
    type: object
...
  models.AchievementTypeStat:
    properties:
      achievementType:
        type: string
      total:
        type: integer
    type: object
...
  models.CompetitionLevelStat:
    properties:
      level:
        type: string
      total:
        type: integer
    type: object
...
  models.LecturerDetailResponse:
    properties:
      department:
        type: string
      fullName:
        description: Diambil dari tabel users
        type: string
      id:
        type: string
      lecturerId:
        type: string
      userId:
        type: string
    type: object
...
  models.StudentAchievementResponse:
    properties:
      achievement_id:
        type: string
      status:
        type: string
      submittedAt:
        type: string
      title:
        type: string
      verifiedAt:
        type: string
    type: object
  models.StudentDetailResponse:
    properties:
      academicYear:
        type: string
      advisorName:
        type: string
      fullName:
        type: string
      id:
        type: string
      programStudy:
        type: string
      studentId:
        type: string
      userId:
        type: string
    type: object
  models.StudentReportResponse:
    properties:
      studentId:
        type: string
      totalPoint:
        type: number
    type: object
  models.TopStudentStat:
    properties:
      fullName:
        type: string
      studentId:
        type: string
      totalPoint:
        type: number
    type: object
  models.UpdateAdvisorRequest:
    properties:
      advisorId:
        type: string
    type: object
...
info:
  contact: {}
  description: API Server untuk manajemen dan pelaporan prestasi mahasiswa.
  title: Sistem Pelaporan Prestasi API
  version: "1.0"
paths:
...
  /api/v1/lecturers:
    get:
      consumes:
      - application/json
      description: Get a list of all lecturers in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LecturerDetailResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List All Lecturers
      tags:
      - Students & Lecturers
  /api/v1/lecturers/{id}/advisees:
    get:
      consumes:
      - application/json
      description: List all students assigned to a specific lecturer
      parameters:
      - description: Lecturer UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentDetailResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Lecturer Advisees
      tags:
      - Students & Lecturers
  /api/v1/reports/statistics:
    get:
      consumes:
      - application/json
      description: 'Retrieve statistics based on role (Admin: Global, Dosen Wali:
        Advisees, Mahasiswa: Own).'
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AchievementStatisticsResponse'
        "403":
          description: Invalid role or access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Achievement Statistics
      tags:
      - Reports
  /api/v1/reports/student/{id}:
    get:
      consumes:
      - application/json
      description: Get total points and student details (Accessible by Mhs for own,
        Dosen for advisees).
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentReportResponse'
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Student Individual Report
      tags:
      - Reports
  /api/v1/students:
    get:
      consumes:
      - application/json
      description: Get a list of all students with their advisors (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentDetailResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List All Students
      tags:
      - Students & Lecturers
  /api/v1/students/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information of a student by their UUID
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentDetailResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Student Detail
      tags:
      - Students & Lecturers
  /api/v1/students/{id}/achievements:
    get:
      consumes:
      - application/json
      description: List all achievement status and points for a specific student
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentAchievementResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Student Achievements
      tags:
      - Students & Lecturers
  /api/v1/students/{id}/advisor:
    put:
      consumes:
      - application/json
      description: Assign a lecturer as an advisor to a student (Admin only)
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      - description: Lecturer ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAdvisorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: advisor updated successfully'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Assign/Update Advisor
      tags:
      - Students & Lecturers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"